+++ app/hooks/useDeleteApi.tsx
@@ -8,22 +8,14 @@ const useDeleteApi = () => {
   const handleDelete = async (modelName: string, idKey: string | number, idValue: string) => {
     setDeleting(true);
     setDeleteResponse(null);
-    console.log(" !! 1.Deleteing ", deleting, " deleteResponse ", JSON.stringify(deleteResponse))
     try {
-      await axios.put(`${process.env.NEXT_PUBLIC_API_URL}/delete`, { modelName, idKey, idValue }); // Note: Axios DELETE requests must send data in the `data` field
-
+      await axios.put(`${process.env.NEXT_PUBLIC_API_URL}/delete`, { modelName, idKey, idValue }); 
       setDeleteResponse({ status: 'success', message: `${modelName} deleted successfully!` });
-      console.log(" !! 2.Deleteing ", deleting, " deleteResponse ", JSON.stringify(deleteResponse))
-
     } catch (error) {
       console.error(`Error deleting ${modelName}:`, error);
       setDeleteResponse({ status: 'error', message: `Failed to delete ${modelName}. Please try again.` });
-      console.log("!! 3.Deleteing ", deleting, " deleteResponse ", JSON.stringify(deleteResponse))
-
     } finally {
       setDeleting(false);
-      console.log("!! 4.Deleteing ", deleting, " deleteResponse ", JSON.stringify(deleteResponse))
-
     }
   };
