--- app/page.tsx
@@ -1,10 +1,0 @@
-'use client'
-import React from 'react';
-import ManageClients from './master-admin/manage/page';
-const RegisterForm: React.FC = () => {
-return (
-    <ManageClients />
-  );
-};
-
-export default RegisterForm;
