--- app/user/page.tsx
@@ -1,9 +1,0 @@
-'use client'
-import React from 'react';
-const RegisterForm: React.FC = () => {
-return (
-    <h1>Hello</h1>
-  );
-};
-
-export default RegisterForm;
