+++ tsconfig.json
@@ -31,6 +31,6 @@
     ".next/types/**/*.ts",
     "app/lib/placeholder-data.js",
     "scripts/seed.js"
-, "app/ui/EmailList.tsx", "app/master-admin/users/MemberTable.jsx", "app/components/master-admin/ItConfig/MailboxConfig.tsx", "app/components/it-admin/CompanyUsersTable.jsx", "app/components/it-admin/FilterDropdown.tsx", "app/api/post/route.ts"  ],
+, "app/ui/EmailList.tsx", "app/admin/users/MemberTable.jsx", "app/components/MasterAdmin/ItConfig/MailboxConfig.tsx", "app/components/UserAdmin/CompanyUsersTable.jsx", "app/components/UserAdmin/FilterDropdown.tsx", "app/api/(crud)/post/route.ts"  ],
   "exclude": ["node_modules"]
 }
